backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
python-dotenv==1.0.1
Pillow==10.4.0
django-redis==5.4.0
gunicorn
whitenoise
django-cors-headers